+++ src/main/kotlin/infrastructure/aggregator/onegamehub/handler/dto/OneGamehubBetDto.kt
@@ -1,5 +1,7 @@
 package com.nekgamebling.infrastructure.aggregator.onegamehub.handler.dto
 
+import java.math.BigInteger
+
 data class OneGameHubBetDto(
     val gameSymbol: String,
 
@@ -9,5 +11,5 @@ data class OneGameHubBetDto(
 
     val freeSpinId: String?,
 
-    val amount: Int
+    val amount: BigInteger
 )
